agent/requirements.txt
ansible==2.2.0.0
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
enum34==1.1.6 ; python_version < '3'
idna==2.7
ipaddress==1.0.22 ; python_version < '3'
jinja2==2.10
markupsafe==1.1.0
paramiko==2.4.2
pyasn1==0.4.4
pycparser==2.19
pycrypto==2.6.1
pynacl==1.3.0
pyyaml==3.13
requests==2.20.1
six==1.11.0
urllib3==1.24.1
